requirements.txt
pyopencl
opencv-python
pillow